added BLMGR

+++ BLMGR_CFG.c
@@ -6,7 +6,7 @@
  */ 
 #include "DIO.h"
 #include "BLMGR_CFG.h"
-BLMGR_DioPinConfig BuzzerConfig      = {PC,1<<4};
-BLMGR_DioPinConfig BlueToothPwrConfig = {PC,1<<5};
-BLMGR_DioPinConfig BluetoothKeyConfig = {PD,1<<2};	
+BLMGR_DioPinConfig BuzzerConfig      = {PC,(u8)1<<4};
+BLMGR_DioPinConfig BlueToothPwrConfig = {PC,(u8)1<<5};
+BLMGR_DioPinConfig BluetoothKeyConfig = {PD,(u8)1<<2};
